requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.8.2
SQLAlchemy==2.0.32
httpx==0.27.0
psycopg2-binary==2.9.9
python-multipart==0.0.9
